✨ feat: BasicUserInfo 컴포넌트 분리 #8
